Redesign Z joints and rename tools
Pointed out by Kayos Maker and totally an issue:

The current design has one screw in contact with the gantry idler
assembly.  This could lead to losing tension in the gantry if you
didn't crank down on the screws, OR if creep over time loosens the
screws to where their positions changes.

This design fixes that issue and also frees up an extra 3 mm of
potential Z motion, however it DOES require the new corners, which
have a 24mm offset.