Update gantry attachments again
Look for 2 new parts marked with _new,
which should work with all released corners.

This is the 4th complete redesign, and hopefully the last!

Should ensure that the idler is rigidly held and never gets loose, even
if the plastic creeps slightly and loosens the screw's grip on the
extrusion under tension.

Note also: the originally-released z_joint_carriage_attachment file is
back; it has not been changed.